Basic Win64 support (MSVC)
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@6439 212acab6-be3b-0410-9dea-997c60f758d6

+++ Contrib/UIs/ui.c
@@ -86,6 +86,6 @@ int WINAPI _tWinMain(HINSTANCE hInst,HINSTANCE hOldInst,LPTSTR CmdLineParams,int
   InitCommonControls();
   g_hInstance = hInst;
   LoadLibrary(_T("RichEd32.dll"));
-  return DialogBox(g_hInstance,MAKEINTRESOURCE(IDD_INST),0,DialogProc);
+  return (int) DialogBox(g_hInstance,MAKEINTRESOURCE(IDD_INST),0,DialogProc);
 }
